requirements.txt
Flask
Pillow
numpy
Werkzeug